requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
distlib==0.3.9
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
filelock==3.17.0
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.6
preshed==3.0.10
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
selenium==4.27.1
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
thinc==8.3.6
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.29.1
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
